requirements.txt
celery 
python-dotenv
redis